+++ src/lib/render/objects/box.hpp
@@ -13,7 +13,7 @@ public:
     box(math::vec3 size) : size_(size) {}
     static std::shared_ptr<box> fromStream(std::stringstream& stream);
 
-    std::optional<long double> intersection(const math::ray& ray) const override;
+    std::optional<float> intersection(const math::ray& ray) const override;
 private:
     math::vec3 size_;
 };
