+++ src/lib/utils/readers.cpp
@@ -3,7 +3,7 @@
 namespace raytracing {
 
 std::istream& operator>>(std::istream& sin, math::vec2& vec) {
-    long double x, y;
+    float x, y;
     sin >> x >> y;
     vec = { x, y };
 
@@ -11,7 +11,7 @@ std::istream& operator>>(std::istream& sin, math::vec2& vec) {
 }
 
 std::istream& operator>>(std::istream& sin, math::vec3& vec) {
-    long double x, y, z;
+    float x, y, z;
     sin >> x >> y >> z;
     vec = { x, y, z };
 
@@ -19,7 +19,7 @@ std::istream& operator>>(std::istream& sin, math::vec3& vec) {
 }
 
 std::istream& operator>>(std::istream& sin, math::vec4& vec) {
-    long double x, y, z, w;
+    float x, y, z, w;
     sin >> x >> y >> z >> w;
     vec = { x, y, z, w };
